+++ container-engine-lib/docs/changelog.md
@@ -1,5 +1,20 @@
 # TBD
 
+# 0.18.0
+### Features
+* Added `ServiceAccounts`, `Roles`, `RoleBindings`, `ClusterRole`, and `ClusterRoleBindings` create and remove methods to `KubernetesManager`
+* Added `CreateEnclave` functionality to Kubernetes backend
+
+### Changes
+* Stopped mounting an enclave data directory on the API container
+
+### Fixes
+* `RunFilesArtifactExpander` now correctly only requires the user to pass in the filepath of the artifact to expand, relative to the enclave data volume root
+
+### Breaking Changes
+* Removed the `enclaveDataDirpathOnHostMachine` parameter from `KurtosisBackend.CreateAPIContainer`
+    * Users no longer need to provide this parameter
+
 # 0.17.0
 ### Features
 * Added `PauseService` and `UnpauseService` to `KurtosisBackend`